Factor out createKeystone
