+++ .changeset/rich-flies-agree.md
@@ -1,0 +1,5 @@
+---
+'@keystone-next/keystone': patch
+---
+
+Factored out `createKeystone` into a separate module.
